move perf test of iterable cursor to the right place
